Smaller api adjustment by user request to reach a final api consencous before beta1 putting all client side apis to lowercase wherever enforced by the local naming conventions
adjusting the demo

moved some stuff from api to impl to get a cleaner api package
optimized the cache handling so that it performs as expected with version on loading resources only once until the cache of the browser itself is dropped!

adjusted the docs to the changed apis where needed!


git-svn-id: https://weblets.dev.java.net/svn/weblets/trunk@178 c71d0f0e-e102-0410-94ad-c789fbfe71d1